Move process_file(s) function out of osmdata.
