+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@ set(osm2pgsql_lib_SOURCES
   expire-tiles.cpp
   gazetteer-style.cpp
   geometry-processor.cpp
+  input.cpp
   logging.cpp
   middle-pgsql.cpp
   middle-ram.cpp
