+++ tests/CMakeLists.txt
@@ -38,6 +38,7 @@ add_definitions(-DOSM2PGSQLDATA_DIR=\"${osm2pgsql_SOURCE_DIR}/\")
 
 add_library(catch_main_lib STATIC catch-main.cpp)
 
+set_test(test-check-input)
 set_test(test-db-copy-thread)
 set_test(test-db-copy-mgr)
 set_test(test-domain-matcher LABELS NoDB)
@@ -46,7 +47,6 @@ set_test(test-middle)
 set_test(test-options-database LABELS NoDB)
 set_test(test-options-parse LABELS NoDB)
 set_test(test-options-projection)
-set_test(test-osmdata)
 set_test(test-output-gazetteer)
 set_test(test-output-multi-line)
 set_test(test-output-multi-point)
